chore: move proto util to proper module
